Add auth to creativity

+++ convex/intentions.ts
@@ -1,10 +1,27 @@
 import moment from "moment-timezone";
 
-import { query, mutation } from "./_generated/server";
 import { v } from "convex/values";
+import { mutation, query } from "./_generated/server";
 
 export const list = query(async (ctx) => {
-  return await ctx.db.query("intentions").collect();
+  const identity = await ctx.auth.getUserIdentity();
+  if (!identity) {
+    throw new Error("Unauthenticated call to mutation");
+  }
+  const user = await ctx.db
+    .query("users")
+    .withIndex("by_token", (q) =>
+      q.eq("tokenIdentifier", identity.tokenIdentifier),
+    )
+    .unique();
+  if (!user) {
+    throw new Error("Unauthenticated call to mutation");
+  }
+
+  return await ctx.db
+    .query("intentions")
+    .filter((q) => q.eq(q.field("userId"), user._id))
+    .collect();
 });
 
 export const get = query({
@@ -12,9 +29,21 @@ export const get = query({
     intentionId: v.id("intentions"),
   },
   async handler(ctx, { intentionId }) {
-    const intention = await ctx.db.get(intentionId);
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("Unauthenticated call to mutation");
+    }
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) =>
+        q.eq("tokenIdentifier", identity.tokenIdentifier),
+      )
+      .unique();
+    if (!user) {
+      throw new Error("Unauthenticated call to mutation");
+    }
 
-    return intention;
+    return await ctx.db.get(intentionId);
   },
 });
 
@@ -28,6 +57,19 @@ export const getByStatus = query({
     ),
   },
   async handler(ctx, { status }) {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("Unauthenticated call to mutation");
+    }
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) =>
+        q.eq("tokenIdentifier", identity.tokenIdentifier),
+      )
+      .unique();
+    if (!user) {
+      throw new Error("Unauthenticated call to mutation");
+    }
     const intentions =
       (await ctx.db
         .query("intentions")
@@ -58,6 +100,7 @@ export const create = mutation({
     if (!user) {
       throw new Error("Unauthenticated call to mutation");
     }
+
     if (title === undefined) {
       title = "";
     }
@@ -67,6 +110,7 @@ export const create = mutation({
       status: "draft",
       emotionId,
       notes,
+      updated: Date.now(),
       userId: user._id,
     });
   },
@@ -77,6 +121,19 @@ export const remove = mutation({
     id: v.id("intentions"),
   },
   async handler(ctx, { id }) {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("Unauthenticated");
+    }
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) =>
+        q.eq("tokenIdentifier", identity.tokenIdentifier),
+      )
+      .unique();
+    if (!user) {
+      throw new Error("Unauthenticated call to mutation");
+    }
     await ctx.db.delete(id);
   },
 });
@@ -97,6 +154,19 @@ export const update = mutation({
     notes: v.optional(v.string()),
   },
   async handler(ctx, { id, title, status, emotionId, notes }) {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("Unauthenticated");
+    }
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) =>
+        q.eq("tokenIdentifier", identity.tokenIdentifier),
+      )
+      .unique();
+    if (!user) {
+      throw new Error("Unauthenticated call to mutation");
+    }
     try {
       const existingIntention = await ctx.db.get(id);
 
@@ -106,10 +176,10 @@ export const update = mutation({
         status: status ?? existingIntention?.status ?? "draft",
         emotionId: emotionId ?? existingIntention?.emotionId,
         notes: notes ?? existingIntention?.notes,
-        updatedAt:
-          status && status !== existingIntention?.status // Currently updatedAt is only updated when the status changes
+        updated:
+          status && status !== existingIntention?.status
             ? moment().tz("America/Denver").valueOf()
-            : existingIntention?.updatedAt ?? undefined,
+            : existingIntention?.updated ?? undefined,
       };
 
       await ctx.db.patch(id, updatedIntention);
